+++ MoMoney/Views/Settings/AdminPage.xaml.cs
@@ -1,0 +1,12 @@
+using MoMoney.ViewModels.Settings;
+
+namespace MoMoney.Views.Settings;
+
+public partial class AdminPage : ContentPage
+{
+	public AdminPage(AdminViewModel vm)
+	{
+		InitializeComponent();
+		BindingContext = vm;
+	}
+}