+++ MoMoney/Views/Settings/ImportExportPage.xaml.cs
@@ -1,0 +1,12 @@
+using MoMoney.ViewModels.Settings;
+
+namespace MoMoney.Views.Settings;
+
+public partial class ImportExportPage : ContentPage
+{
+	public ImportExportPage(ImportExportViewModel vm)
+	{
+		InitializeComponent();
+		BindingContext = vm;
+	}
+}